src/steam_catalogue_cloud_requirements.txt
httpx==0.24.1
avro-python3==1.10.2
google-cloud-storage==2.10.0
functions-framework==3.3.0
httpx==0.24.1